243: move LogEntry
